pyproject.toml
[tool.poetry]
name = "quantuminspire"
version = "0.1.0"
description = "SDK for the Quantum Inspire platform."
authors = ["Quantum Inspire <support@quantum-inspire.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://www.quantum-inspire.com"
repository = "https://github.com/quantum-inspire-2"
classifiers = [
  'Development Status :: 4 - Beta',
  'Programming Language :: Python :: 3',
  'Programming Language :: Python :: 3.8',
  'Programming Language :: Python :: 3.9',
  'Programming Language :: Python :: 3.10',
  'Programming Language :: Python :: 3.11',
  'License :: OSI Approved :: Apache Software License'
]

[tool.poetry.scripts]
qi = "quantuminspire.cli.command_list:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.12.3"}
qutechopenql = "^0.12.2"
pydantic = "^2.7.0"
qi-compute-api-client = {git = "https://github.com/QuTech-Delft/compute-api-client.git", branch="0.27.0"}
qxelarator = {version = "^0.6.2", optional = true}
pydantic-settings = "^2.2.1"
qiskit = "^0.45.2"
qiskit-algorithms = "^0.3.0"
qiskit-nature = "^0.7.1"
pyscf = {version = "^2.4.0", markers = 'sys_platform != "win32"'}
oauthlib = "^3.2.2"
requests = "^2.31.0"


[tool.poetry.extras]
local = ["qxelarator"]

[tool.poetry.group.dev.dependencies]
pytest = {extras = ["toml"], version = "^8.1.1"}
pytest-cov = "^5.0.0"
coverage = {extras = ["toml"], version = "^7.4.4"}
mypy = "^1.9"
docformatter = "1.7.5"
pytest-mock = "^3.14.0"
black = "^24.3.0"
pytest-asyncio = "^0.23.6"
ruff = "^0.3.5"
types-requests = "^2.25.0"
types-oauthlib = "^3.2.0.20240217"
responses = "^0.25.0"
freezegun = "^1.4.0"


[tool.pytest.ini_options]
addopts = "-v --cov --cov-report term-missing:skip-covered --cov-report xml"
asyncio_mode = "auto"
testpaths = ["tests"]


[tool.coverage.run]
branch = true
source = ["quantuminspire"]
omit = ["tests/*", "**/.tox/**", "scratch"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]

[tool.black]
line-length = 120

[tool.mypy]
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
ignore_missing_imports = true
show_error_codes = true
plugins = [
    "pydantic.mypy"
]
exclude = [
    "examples/",
    "tests/sdk/qiskit/",
    "scratch"
]

[tool.ruff]
lint.select = ["E", "F", "I"]
line-length = 120
target-version = "py311"
lint.isort.combine-as-imports = true

[tool.pylint]
MASTER.extension-pkg-whitelist = "pydantic"
FORMAT.max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.docformatter]
recursive = true
wrap-summaries = 119
wrap-descriptions = 119

[tool.pydantic-mypy]
warn_untyped_fields = true
